+++ webots_ros2_mavic/launch/robot_launch.py
@@ -17,16 +17,14 @@
 """Launch Webots Mavic 2 Pro driver."""
 
 import os
-import pathlib
 import launch
 from launch.substitutions import LaunchConfiguration
 from launch.actions import DeclareLaunchArgument
 from launch.substitutions.path_join_substitution import PathJoinSubstitution
-from launch_ros.actions import Node
 from launch import LaunchDescription
 from ament_index_python.packages import get_package_share_directory
 from webots_ros2_driver.webots_launcher import WebotsLauncher
-from webots_ros2_driver.utils import controller_url_prefix
+from webots_ros2_driver.webots_controller import WebotsController
 
 
 def generate_launch_description():
@@ -38,14 +36,11 @@ def generate_launch_description():
         ros2_supervisor=True
     )
 
-    robot_description = pathlib.Path(os.path.join(package_dir, 'resource', 'mavic_webots.urdf')).read_text()
-    mavic_driver = Node(
-        package='webots_ros2_driver',
-        executable='driver',
-        output='screen',
-        additional_env={'WEBOTS_CONTROLLER_URL': controller_url_prefix() + 'Mavic_2_PRO'},
+    robot_description_path = os.path.join(package_dir, 'resource', 'mavic_webots.urdf')
+    mavic_driver = WebotsController(
+        robot_name='Mavic_2_PRO',
         parameters=[
-            {'robot_description': robot_description},
+            {'robot_description': robot_description_path},
         ],
         respawn=True
     )